% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_databases.R
\name{describe_databases}
\alias{describe_databases}
\title{Describe eulaw.app databases}
\usage{
describe_databases()
}
\value{
A tibble with 3 columns.
}
\description{
This function describes all of the databases that are available via the eulaw.app
API.
}
