% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datasets.R
\name{describe_datasets}
\alias{describe_datasets}
\title{Describes datasets in an eulaw.app database}
\usage{
describe_datasets(database)
}
\arguments{
\item{database}{The name of the database as a string. Use the function
\code{list_databases()} to see a list of valid values.}
}
\value{
A tibble with 3 columns.
}
\description{
This function describes all of the datasets that are available from the
requested database.
}
