% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_databases.R
\name{list_databases}
\alias{list_databases}
\title{List eulaw.app databases}
\usage{
list_databases()
}
\value{
A tibble with 2 columns.
}
\description{
This function lists all of the databases that are available via the eulaw.app
API.
}
