% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets in an eulaw.app database}
\usage{
list_datasets(database)
}
\arguments{
\item{database}{The name of the database as a string. Use the function
\code{list_databases()} to see a list of valid values.}
}
\value{
A tibble with 2 columns.
}
\description{
This function lists all of the datasets that are available from the requested
database.
}
