% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_parameter_values.R
\name{list_parameter_values}
\alias{list_parameter_values}
\title{List values for an eulaw.app API parameter}
\usage{
list_parameter_values(database, parameter)
}
\arguments{
\item{database}{The name of the database as a string. Use the function
\code{list_databases()} to see a list of valid values.}

\item{parameter}{The name of an API parameter as a string. Use the function
\code{list_parameters()} to see a list of valid values.}
}
\value{
A tibble with 2 columns.
}
\description{
This function lists all of the valid values for an API parameter. API
parameters can appear across multiple datasets within a database, and they
are always coded the same within a database, so you only need to specify a
database and a parameter. API parameters take numeric values that correspond
to categorical values.
}
