% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_parameters.R
\name{list_parameters}
\alias{list_parameters}
\title{List API parameters for an eulaw.app dataset}
\usage{
list_parameters(database, dataset)
}
\arguments{
\item{database}{The name of the database as a string. Use the function
\code{list_databases()} to see a list of valid values.}

\item{dataset}{The name of the dataset as a string. Use the function
\code{list_datasets()} to see a list of valid values.}
}
\value{
A tibble with 2 columns.
}
\description{
This function lists all of the valid API parameters for the requested
dataset. You can use API parameters to filter the data when downloading a
dataset via the eulaw.app API. See the function \code{download_data()} for
more details. See the package vignette for an example of how to specify API
parameters.
}
