% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variables.R
\name{list_variables}
\alias{list_variables}
\title{List variables in an eulaw.app dataset}
\usage{
list_variables(database, dataset)
}
\arguments{
\item{database}{The name of the database as a string. Use the function
\code{list_databases()} to see a list of valid values.}

\item{dataset}{The name of the dataset as a string. Use the function
\code{list_datasets()} to see a list of valid values.}
}
\value{
A tibble with 2 columns.
}
\description{
This function lists all of the variables that are available in the requested
dataset.
}
